# swagger.yaml
openapi: 3.0.3
info:
  title: Smart Task & Productivity Management System API
  description: |
    Full-Stack Task Management System with Role-Based Access Control,
    JWT Authentication, and AI Productivity Insights.

    ## Authentication
    All endpoints (except `/api/auth/*`) require a **Bearer JWT token**.
    1. Call `POST /api/auth/login` to get a token.
    2. Click **Authorize** (lock icon) and enter: `Bearer <your_token>`

    ## Roles
    | Role | Permissions |
    |------|-------------|
    | **MANAGER** | Create/assign/delete any task, view all users, full analytics |
    | **EMPLOYEE** | View/update only their assigned tasks, personal analytics |

    ## Base URL
    `http://localhost:8081`

    ## Demo Credentials
    - Manager: `manager@taskmanager.com` / `manager123`
    - Employee: `priya@employee.com` / `employee123`
  version: 1.0.0
  contact:
    name: Task Management Team
    email: admin@taskmanager.com
  license:
    name: MIT License

servers:
  - url: http://localhost:8081
    description: Local development server

tags:
  - name: Authentication
    description: Register and login — no token required
  - name: Tasks
    description: Task CRUD — role-filtered (Manager sees all, Employee sees assigned)
  - name: Categories
    description: Task category management
  - name: Users
    description: User listing — Manager only
  - name: Analytics
    description: Dashboard metrics and productivity score
  - name: Insights
    description: AI-powered productivity insights

# ─── Security Scheme ──────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from POST /api/auth/login

  schemas:
    # ── Wrapper ───────────────────────────────────────────────────────────────
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          description: Response payload (type varies per endpoint)

    # ── Auth ──────────────────────────────────────────────────────────────────
    RegisterRequest:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          example: "Arjun Sharma"
        email:
          type: string
          format: email
          example: "arjun@company.com"
        password:
          type: string
          minLength: 6
          example: "secret123"
        role:
          type: string
          enum: [MANAGER, EMPLOYEE]
          example: "EMPLOYEE"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "manager@taskmanager.com"
        password:
          type: string
          example: "manager123"

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        email:
          type: string
          example: "manager@taskmanager.com"
        name:
          type: string
          example: "Arjun Sharma"
        role:
          type: string
          enum: [MANAGER, EMPLOYEE]
          example: "MANAGER"
        userId:
          type: integer
          format: int64
          example: 1

    # ── Task ──────────────────────────────────────────────────────────────────
    TaskDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
          readOnly: true
        title:
          type: string
          example: "Implement JWT Authentication"
        description:
          type: string
          nullable: true
          example: "Set up Spring Security with JWT tokens."
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, OVERDUE]
          example: "IN_PROGRESS"
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
          example: "HIGH"
        dueDate:
          type: string
          format: date
          nullable: true
          example: "2026-05-20"
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: "2026-05-08T10:30:00"
        assignedToId:
          type: integer
          format: int64
          nullable: true
          example: 2
        assignedToName:
          type: string
          nullable: true
          readOnly: true
          example: "Priya Patel"
        createdById:
          type: integer
          format: int64
          nullable: true
          readOnly: true
          example: 1
        createdByName:
          type: string
          nullable: true
          readOnly: true
          example: "Arjun Sharma"
        categoryId:
          type: integer
          format: int64
          nullable: true
          example: 1
        categoryName:
          type: string
          nullable: true
          readOnly: true
          example: "Development"

    TaskCreateRequest:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
          example: "Build Dashboard Charts"
        description:
          type: string
          example: "Integrate Chart.js for analytics visualization."
        status:
          type: string
          enum: [PENDING, IN_PROGRESS, COMPLETED, OVERDUE]
          example: "PENDING"
        priority:
          type: string
          enum: [LOW, MEDIUM, HIGH]
          example: "MEDIUM"
        dueDate:
          type: string
          format: date
          nullable: true
          example: "2026-05-25"
        assignedToId:
          type: integer
          format: int64
          nullable: true
          description: Employee ID to assign to. If omitted, assigned to creator.
          example: 2
        categoryId:
          type: integer
          format: int64
          nullable: true
          example: 1

    # ── Category ──────────────────────────────────────────────────────────────
    CategoryDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          example: 1
        categoryName:
          type: string
          example: "Development"

    CategoryCreateRequest:
      type: object
      required: [categoryName]
      properties:
        categoryName:
          type: string
          example: "DevOps"

    # ── User ──────────────────────────────────────────────────────────────────
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 2
        name:
          type: string
          example: "Priya Patel"
        email:
          type: string
          format: email
          example: "priya@employee.com"
        role:
          type: string
          enum: [MANAGER, EMPLOYEE]
          example: "EMPLOYEE"

    # ── Analytics ─────────────────────────────────────────────────────────────
    AnalyticsDTO:
      type: object
      properties:
        totalTasks:
          type: integer
          example: 10
        completedTasks:
          type: integer
          example: 3
        pendingTasks:
          type: integer
          example: 4
        inProgressTasks:
          type: integer
          example: 2
        overdueTasks:
          type: integer
          example: 1
        highPriorityTasks:
          type: integer
          example: 4
        mediumPriorityTasks:
          type: integer
          example: 4
        lowPriorityTasks:
          type: integer
          example: 2
        completionRate:
          type: number
          format: double
          example: 30.0
        productivityScore:
          type: integer
          minimum: 0
          maximum: 100
          example: 51
        tasksByStatus:
          type: object
          additionalProperties:
            type: integer
          example:
            PENDING: 4
            IN_PROGRESS: 2
            COMPLETED: 3
            OVERDUE: 1
        tasksByPriority:
          type: object
          additionalProperties:
            type: integer
          example:
            LOW: 2
            MEDIUM: 4
            HIGH: 4

    # ── Insights ──────────────────────────────────────────────────────────────
    InsightDTO:
      type: object
      properties:
        insights:
          type: array
          items:
            type: string
          example:
            - "You have 1 overdue task that needs immediate attention."
            - "4 high-priority tasks require focus — prioritize these first."
            - "Your completion rate is 30% — there is room for improvement."
        overallStatus:
          type: string
          enum: [EXCELLENT, ON_TRACK, AT_RISK, CRITICAL, NO_TASKS]
          example: "AT_RISK"
        productivityScore:
          type: integer
          example: 51
        recommendation:
          type: string
          example: "Focus on completing in-progress tasks before taking on new ones."

    # ── Error ─────────────────────────────────────────────────────────────────
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Access denied"
        data:
          nullable: true
          example: null

# ─── Global security (all endpoints except auth) ─────────────────────────────
security:
  - bearerAuth: []

# ─── Paths ───────────────────────────────────────────────────────────────────
paths:

  # ════════════════════════════════════════════════════════
  #  AUTHENTICATION  (no token required)
  # ════════════════════════════════════════════════════════

  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Creates a new MANAGER or EMPLOYEE account and returns a JWT token.
        No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              manager:
                summary: Register as Manager
                value:
                  name: "Arjun Sharma"
                  email: "arjun@company.com"
                  password: "secret123"
                  role: "MANAGER"
              employee:
                summary: Register as Employee
                value:
                  name: "Priya Patel"
                  email: "priya@company.com"
                  password: "secret123"
                  role: "EMPLOYEE"
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthResponse'
              example:
                success: true
                message: "User registered successfully"
                data:
                  token: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJhcmp1..."
                  email: "arjun@company.com"
                  name: "Arjun Sharma"
                  role: "MANAGER"
                  userId: 5
        '400':
          description: Email already registered or validation error
          content:
            application/json:
              example:
                success: false
                message: "Email already registered: arjun@company.com"
                data: null

  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Login and get JWT token
      description: |
        Authenticates with email + password and returns a JWT token valid for **24 hours**.
        Use this token as `Authorization: Bearer <token>` on all protected endpoints.
        No authentication required.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              manager:
                summary: Manager login
                value:
                  email: "manager@taskmanager.com"
                  password: "manager123"
              employee:
                summary: Employee login
                value:
                  email: "priya@employee.com"
                  password: "employee123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              example:
                success: true
                message: "Login successful"
                data:
                  token: "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJtYW5h..."
                  email: "manager@taskmanager.com"
                  name: "Arjun Sharma"
                  role: "MANAGER"
                  userId: 1
        '401':
          description: Invalid credentials
          content:
            application/json:
              example:
                success: false
                message: "Bad credentials"
                data: null

  # ════════════════════════════════════════════════════════
  #  TASKS
  # ════════════════════════════════════════════════════════

  /api/tasks:
    get:
      tags: [Tasks]
      summary: Get all tasks
      description: |
        - **MANAGER**: returns every task in the system
        - **EMPLOYEE**: returns only tasks assigned to them

        Tasks past their due date and not completed are automatically marked OVERDUE.
      responses:
        '200':
          description: Tasks retrieved
          content:
            application/json:
              example:
                success: true
                message: "Tasks retrieved"
                data:
                  - id: 1
                    title: "Design Login UI"
                    description: "Create wireframes and Figma designs for the login page."
                    status: "COMPLETED"
                    priority: "HIGH"
                    dueDate: "2026-05-06"
                    createdAt: "2026-05-01T09:00:00"
                    assignedToId: 2
                    assignedToName: "Priya Patel"
                    createdById: 1
                    createdByName: "Arjun Sharma"
                    categoryId: 2
                    categoryName: "Design"
                  - id: 3
                    title: "Build Dashboard Charts"
                    description: "Integrate Chart.js for analytics visualization."
                    status: "IN_PROGRESS"
                    priority: "MEDIUM"
                    dueDate: "2026-05-10"
                    createdAt: "2026-05-01T09:05:00"
                    assignedToId: 2
                    assignedToName: "Priya Patel"
                    createdById: 1
                    createdByName: "Arjun Sharma"
                    categoryId: 1
                    categoryName: "Development"
        '401':
          description: Missing or invalid JWT token

    post:
      tags: [Tasks]
      summary: Create a new task
      description: |
        Creates a new task.
        - **MANAGER**: can assign to any employee via `assignedToId`
        - **EMPLOYEE**: task is auto-assigned to themselves; `assignedToId` is ignored
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
            examples:
              manager_creates:
                summary: Manager assigns task to employee
                value:
                  title: "Setup CI/CD Pipeline"
                  description: "Configure GitHub Actions for automated deployment."
                  status: "PENDING"
                  priority: "HIGH"
                  dueDate: "2026-05-30"
                  assignedToId: 3
                  categoryId: 5
              employee_creates:
                summary: Employee creates personal task
                value:
                  title: "Review pull requests"
                  description: "Review open PRs before end of day."
                  status: "PENDING"
                  priority: "MEDIUM"
                  dueDate: "2026-05-10"
                  assignedToId: null
                  categoryId: null
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              example:
                success: true
                message: "Task created successfully"
                data:
                  id: 11
                  title: "Setup CI/CD Pipeline"
                  description: "Configure GitHub Actions for automated deployment."
                  status: "PENDING"
                  priority: "HIGH"
                  dueDate: "2026-05-30"
                  createdAt: "2026-05-09T11:00:00"
                  assignedToId: 3
                  assignedToName: "Sneha Gupta"
                  createdById: 1
                  createdByName: "Arjun Sharma"
                  categoryId: 5
                  categoryName: "DevOps"
        '400':
          description: Validation error (title missing, invalid status/priority)
        '401':
          description: Missing or invalid JWT token

  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task by ID
      description: |
        - **MANAGER**: can access any task
        - **EMPLOYEE**: can only access tasks assigned to them (403 otherwise)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 3
      responses:
        '200':
          description: Task retrieved
          content:
            application/json:
              example:
                success: true
                message: "Task retrieved"
                data:
                  id: 3
                  title: "Build Dashboard Charts"
                  status: "IN_PROGRESS"
                  priority: "MEDIUM"
                  dueDate: "2026-05-10"
                  assignedToName: "Priya Patel"
                  categoryName: "Development"
        '403':
          description: Employee accessing another user's task
        '404':
          description: Task not found

    put:
      tags: [Tasks]
      summary: Update an existing task
      description: |
        Full update of a task.
        - **MANAGER**: can update any task and reassign to a different employee
        - **EMPLOYEE**: can only update their own tasks (status, priority, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
            examples:
              employee_marks_done:
                summary: Employee marks task as completed
                value:
                  title: "Build Dashboard Charts"
                  description: "Integrate Chart.js for analytics visualization."
                  status: "COMPLETED"
                  priority: "MEDIUM"
                  dueDate: "2026-05-10"
                  assignedToId: 2
                  categoryId: 1
              manager_reassigns:
                summary: Manager updates and reassigns task
                value:
                  title: "Build Dashboard Charts"
                  description: "Integrate Chart.js for analytics visualization."
                  status: "IN_PROGRESS"
                  priority: "HIGH"
                  dueDate: "2026-05-15"
                  assignedToId: 3
                  categoryId: 1
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              example:
                success: true
                message: "Task updated"
                data:
                  id: 3
                  title: "Build Dashboard Charts"
                  status: "COMPLETED"
                  priority: "MEDIUM"
                  dueDate: "2026-05-10"
                  assignedToName: "Priya Patel"
        '401':
          description: Missing or invalid JWT token
        '403':
          description: Employee trying to update another user's task
        '404':
          description: Task not found

    delete:
      tags: [Tasks]
      summary: Delete a task
      description: |
        Permanently deletes a task.
        **MANAGER only** — employees receive 403.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 3
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              example:
                success: true
                message: "Task deleted"
                data: null
        '403':
          description: Employee cannot delete tasks
        '404':
          description: Task not found

  # ════════════════════════════════════════════════════════
  #  CATEGORIES
  # ════════════════════════════════════════════════════════

  /api/categories:
    get:
      tags: [Categories]
      summary: Get all categories
      description: Returns all task categories. Available to both roles.
      responses:
        '200':
          description: Categories retrieved
          content:
            application/json:
              example:
                success: true
                message: "Categories retrieved"
                data:
                  - id: 1
                    categoryName: "Development"
                  - id: 2
                    categoryName: "Design"
                  - id: 3
                    categoryName: "Marketing"
                  - id: 4
                    categoryName: "QA Testing"
                  - id: 5
                    categoryName: "DevOps"

    post:
      tags: [Categories]
      summary: Create a new category
      description: Creates a new task category. Available to both roles.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreateRequest'
            example:
              categoryName: "Machine Learning"
      responses:
        '200':
          description: Category created
          content:
            application/json:
              example:
                success: true
                message: "Category created"
                data:
                  id: 6
                  categoryName: "Machine Learning"
        '400':
          description: Validation error

  /api/categories/{id}:
    delete:
      tags: [Categories]
      summary: Delete a category
      description: Deletes a category by ID. Available to both roles.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 6
      responses:
        '200':
          description: Category deleted
          content:
            application/json:
              example:
                success: true
                message: "Category deleted"
                data: null
        '404':
          description: Category not found

  # ════════════════════════════════════════════════════════
  #  USERS  (Manager only)
  # ════════════════════════════════════════════════════════

  /api/users:
    get:
      tags: [Users]
      summary: Get all users
      description: Returns all users (both managers and employees). **MANAGER only.**
      responses:
        '200':
          description: Users retrieved
          content:
            application/json:
              example:
                success: true
                message: "Users retrieved"
                data:
                  - id: 1
                    name: "Arjun Sharma"
                    email: "manager@taskmanager.com"
                    role: "MANAGER"
                  - id: 2
                    name: "Priya Patel"
                    email: "priya@employee.com"
                    role: "EMPLOYEE"
        '403':
          description: Employee access denied

  /api/users/employees:
    get:
      tags: [Users]
      summary: Get all employees
      description: |
        Returns only EMPLOYEE-role users.
        Used by the task form to populate the "Assign To" dropdown.
        **MANAGER only.**
      responses:
        '200':
          description: Employees retrieved
          content:
            application/json:
              example:
                success: true
                message: "Employees retrieved"
                data:
                  - id: 2
                    name: "Priya Patel"
                    email: "priya@employee.com"
                    role: "EMPLOYEE"
                  - id: 3
                    name: "Rahul Verma"
                    email: "rahul@employee.com"
                    role: "EMPLOYEE"
                  - id: 4
                    name: "Sneha Gupta"
                    email: "sneha@employee.com"
                    role: "EMPLOYEE"
        '403':
          description: Employee access denied

  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by ID
      description: Returns a single user by their ID. Available to both roles.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 2
      responses:
        '200':
          description: User retrieved
          content:
            application/json:
              example:
                success: true
                message: "User retrieved"
                data:
                  id: 2
                  name: "Priya Patel"
                  email: "priya@employee.com"
                  role: "EMPLOYEE"
        '404':
          description: User not found

  # ════════════════════════════════════════════════════════
  #  ANALYTICS
  # ════════════════════════════════════════════════════════

  /api/analytics:
    get:
      tags: [Analytics]
      summary: Get dashboard analytics
      description: |
        Returns task statistics and productivity metrics.
        - **MANAGER**: team-wide analytics (all tasks)
        - **EMPLOYEE**: personal analytics (only their tasks)
      responses:
        '200':
          description: Analytics retrieved
          content:
            application/json:
              example:
                success: true
                message: "Analytics retrieved"
                data:
                  totalTasks: 10
                  completedTasks: 2
                  pendingTasks: 4
                  inProgressTasks: 3
                  overdueTasks: 1
                  highPriorityTasks: 4
                  mediumPriorityTasks: 4
                  lowPriorityTasks: 2
                  completionRate: 20.0
                  productivityScore: 44
                  tasksByStatus:
                    PENDING: 4
                    IN_PROGRESS: 3
                    COMPLETED: 2
                    OVERDUE: 1
                  tasksByPriority:
                    LOW: 2
                    MEDIUM: 4
                    HIGH: 4
        '401':
          description: Missing or invalid JWT token

  # ════════════════════════════════════════════════════════
  #  INSIGHTS
  # ════════════════════════════════════════════════════════

  /api/insights:
    get:
      tags: [Insights]
      summary: Get AI productivity insights
      description: |
        Returns rule-based AI insights generated from the user's task data.
        Includes a list of insight messages, an overall status, productivity score,
        and a key recommendation.
        - **MANAGER**: insights based on full team task data
        - **EMPLOYEE**: insights based on their assigned tasks
      responses:
        '200':
          description: Insights generated
          content:
            application/json:
              examples:
                at_risk:
                  summary: AT_RISK status example
                  value:
                    success: true
                    message: "Insights generated"
                    data:
                      insights:
                        - "⚠️ You have 1 overdue task that needs immediate attention."
                        - "🔴 4 high-priority tasks require your focus — tackle these first."
                        - "📊 Your completion rate is 20% — there is significant room for improvement."
                        - "⏳ 3 tasks are currently in progress. Stay focused to complete them."
                      overallStatus: "AT_RISK"
                      productivityScore: 44
                      recommendation: "Focus on resolving overdue tasks immediately, then work through high-priority items systematically."
                excellent:
                  summary: EXCELLENT status example
                  value:
                    success: true
                    message: "Insights generated"
                    data:
                      insights:
                        - "🌟 Outstanding performance! You have completed all your tasks."
                        - "✅ Keep up the excellent work and take on new challenges."
                      overallStatus: "EXCELLENT"
                      productivityScore: 100
                      recommendation: "Excellent work! Consider mentoring teammates or taking on stretch goals."
        '401':
          description: Missing or invalid JWT token