# task-management-backend/src/main/java/com/taskmanager/dto/UserDTO.java
package com.taskmanager.dto;

import com.taskmanager.entity.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/** DTO for returning user information (no password exposed) */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserDTO {
    private Long id;
    private String name;
    private String email;
    private Role role;
}